CAMEL-17026: initial refactoring to allow easier implementation of the feature
